+++ frontend/src/common/mock/mentoring/handlers.ts
@@ -58,18 +58,4 @@ const getMentoringItems = http.get(MENTORING_URL, ({ request }) => {
   }
 });
 
-const postMentoringCreate = http.post(MENTORING_URL, async ({ request }) => {
-  const formData = await request.formData();
-
-  const dataJson = formData.get('data');
-
-  const parsedData = JSON.parse(dataJson as string);
-
-  if (!parsedData) {
-    return HttpResponse.json({ message: 'Bad Request' }, { status: 400 });
-  }
-
-  return HttpResponse.json({ message: true }, { status: 201 });
-});
-
-export const mentoringHandler = [getMentoringItems, postMentoringCreate];
+export const mentoringHandler = [getMentoringItems];
